requirements.txt
ebooklib==0.16
lxml==3.8.0
pillow==4.2.1
